+++ react/ProductSummaryQuantity.tsx
@@ -10,7 +10,7 @@ import BaseProductQuantity, {
 const CSS_HANDLES = ['summaryContainer'] as const
 
 const ProductSummaryQuantity: StorefrontFunctionComponent<BaseProps> = props => {
-  const { warningQuantityThreshold, showLabel, size, selectorType } = props
+  const { warningQuantityThreshold, showLabel, size, selectorType, showSuffix } = props
   const handles = useCssHandles(CSS_HANDLES)
   const { selectedItem, selectedQuantity } = useProduct()
   const dispatch = useProductDispatch()
@@ -27,6 +27,7 @@ const ProductSummaryQuantity: StorefrontFunctionComponent<BaseProps> = props => 
       onClick={handleClick}
       className={`${handles.summaryContainer} center mw-100`}>
       <BaseProductQuantity
+        showSuffix={showSuffix}
         size={size}
         dispatch={dispatch}
         showLabel={showLabel}